[FSSDK-10950] cleanup of test files
