+++ lib/event_processor/event_dispatcher/default_dispatcher.node.spec.ts
@@ -20,13 +20,13 @@ vi.mock('./default_dispatcher', () => {
   return { DefaultEventDispatcher };
 });
 
-vi.mock('../../utils/http_request_handler/node_request_handler', () => {
+vi.mock('../../utils/http_request_handler/request_handler.node', () => {
   const NodeRequestHandler = vi.fn();
   return { NodeRequestHandler };
 });
 
 import { DefaultEventDispatcher } from './default_dispatcher';
-import { NodeRequestHandler } from '../../utils/http_request_handler/node_request_handler';
+import { NodeRequestHandler } from '../../utils/http_request_handler/request_handler.node';
 import eventDispatcher from './default_dispatcher.node';
 
 describe('eventDispatcher', () => {
