+++ lib/odp/odp_manager_factory.node.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { NodeRequestHandler } from '../utils/http_request_handler/node_request_handler';
+import { NodeRequestHandler } from '../utils/http_request_handler/request_handler.node';
 import { eventApiRequestGenerator } from './event_manager/odp_event_api_manager';
 import { OdpManager } from './odp_manager';
 import { getOdpManager, OdpManagerOptions } from './odp_manager_factory';
