+++ lib/odp/odp_manager_factory.react_native.spec.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-vi.mock('../utils/http_request_handler/browser_request_handler', () => {
+vi.mock('../utils/http_request_handler/request_handler.browser', () => {
   return { BrowserRequestHandler: vi.fn() };
 });
 
@@ -26,7 +26,7 @@ vi.mock('./odp_manager_factory', () => {
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { getOdpManager, OdpManagerOptions } from './odp_manager_factory';
 import { RN_DEFAULT_API_TIMEOUT, RN_DEFAULT_BATCH_SIZE, RN_DEFAULT_FLUSH_INTERVAL, createOdpManager } from './odp_manager_factory.react_native';
-import { BrowserRequestHandler } from '../utils/http_request_handler/browser_request_handler'
+import { BrowserRequestHandler } from '../utils/http_request_handler/request_handler.browser'
 import { eventApiRequestGenerator } from './event_manager/odp_event_api_manager';
 
 describe('createOdpManager', () => {
