+++ lib/project_config/config_manager_factory.react_native.ts
@@ -15,7 +15,7 @@
  */
 
 import { getPollingConfigManager, PollingConfigManagerConfig } from "./config_manager_factory";
-import { BrowserRequestHandler } from "../utils/http_request_handler/browser_request_handler";
+import { BrowserRequestHandler } from "../utils/http_request_handler/request_handler.browser";
 import { ProjectConfigManager } from "./project_config_manager";
 import { AsyncStorageCache } from "../utils/cache/async_storage_cache.react_native";
