+++ tests/testUtils.ts
@@ -16,46 +16,3 @@
 
 import { vi } from 'vitest';
 
-import PersistentKeyValueCache from "../lib/plugins/key_value_cache/persistentKeyValueCache";
-
-export function advanceTimersByTime(waitMs: number): Promise<void> {
-  const timeoutPromise: Promise<void> = new Promise(res => setTimeout(res, waitMs));
-  vi.advanceTimersByTime(waitMs);
-  return timeoutPromise;
-}
-
-export function getTimerCount(): number {
-  return vi.getTimerCount();
-}
-
-export const getTestPersistentCache = (): PersistentKeyValueCache => {
-  const cache = {
-    get: vi.fn().mockImplementation((key: string): Promise<string | undefined> => {
-      let val : string | undefined = undefined;
-      switch (key) {
-        case 'opt-datafile-keyThatExists':
-          val = JSON.stringify({ name: 'keyThatExists' });
-          break;
-      }
-      return Promise.resolve(val);
-    }),
-
-    set: vi.fn().mockImplementation((): Promise<void> => {
-      return Promise.resolve();
-    }),
-
-    contains: vi.fn().mockImplementation((): Promise<boolean> => {
-      return Promise.resolve(false);
-    }),
-
-    remove: vi.fn().mockImplementation((): Promise<boolean> => {
-      return Promise.resolve(false);
-    }),
-  };
-
-  return cache;
-}
-
-export const wait = (ms: number) => {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-};
